+++ mythic_container/PayloadBuilder.py
@@ -49,6 +49,8 @@ class BuildParameterType(str, Enum):
             A string value
         ChooseOne:
             A list of choices for the user to select exactly one
+        ChooseMultiple:
+            A list of choices for the user to select 0 or more
         Array:
             The user can supply multiple values in an Array format
         Date:
@@ -59,14 +61,18 @@ class BuildParameterType(str, Enum):
             The user can toggle a switch for True/False
         File:
             The user can select a file that gets uploaded - a file UUID gets passed in during build
+        TypedArray:
+            The user can supply an array where each element also has a drop-down option of choices
     """
     String = "String"
     ChooseOne = "ChooseOne"
+    ChooseMultiple = "ChooseMultiple"
     Array = "Array"
     Date = "Date"
     Dictionary = "Dictionary"
     Boolean = "Boolean"
     File = "File"
+    TypedArray = "TypedArray"
 
 
 class DictionaryChoice:
@@ -455,7 +461,9 @@ class PayloadType:
             List of C2 Profiles that the user selected to build into the agent along with their parameter values
         commands (CommandList):
             List of commands the user selected to build into the agent
-        wrapped_payload (str):
+        wrapped_payload (bytes):
+            If this is a wrapper payload type, this is the raw bytes of the payload to wrap
+        wrapped_payload_uuid (str):
             If this is a wrapper payload type, this is the UUID of the payload to wrap
         selected_os (str):
             The OS the user selected for the first step when building this payload
@@ -487,7 +495,8 @@ class PayloadType:
     uuid: str = None
     c2info: [C2ProfileParameters] = None
     commands: CommandList = None
-    wrapped_payload: str = None
+    wrapped_payload_uuid: str = None
+    wrapped_payload: bytes = None
     selected_os: str = None
     filename: str = None
     build_steps = []
@@ -508,7 +517,8 @@ class PayloadType:
             c2info: [C2ProfileParameters] = None,
             selected_os: str = None,
             commands: CommandList = None,
-            wrapped_payload: str = None,
+            wrapped_payload_uuid: str = None,
+            wrapped_payload: bytes = None
     ):
         self.uuid = uuid
         self.c2info = c2info
@@ -516,6 +526,7 @@ class PayloadType:
         self.commands = commands
         self.filename = filename
         self.wrapped_payload = wrapped_payload
+        self.wrapped_payload_uuid = wrapped_payload_uuid
         if self.agent_path is None:
             self.agent_path = pathlib.Path(".") / self.name
             logger.error(f"{self.name} has no agent_path set, setting to {self.agent_path}")